API Football Link with RabbitMQ
Can run command to send a message to RabbitMQ. This will fetch data depending on the status of the current match (if gates are open and therefore if we should be fetching live match data etc.)
The script will then ensure we do not call the API more than once every 20 seconds.
It saves the data to the database. Need to check for issues still.